+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/ast/SourceCodePositioner.java
@@ -54,7 +54,7 @@ public class SourceCodePositioner {
     public int columnFromOffset(int lineNumber, int offset) {
         int lineIndex = lineNumber - 1;
         if (lineIndex < 0 || lineIndex >= lineOffsets.length) {
-            // no line found
+            // no line number found...
             return 0;
         }
         int columnOffset = offset - lineOffsets[lineNumber - 1];