+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symboltable/ClassScopeTest.java
@@ -12,21 +12,20 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
+import org.junit.Test;
+
 import net.sourceforge.pmd.PMD;
 import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
 import net.sourceforge.pmd.lang.java.ast.DummyJavaNode;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
-import net.sourceforge.pmd.lang.java.symboltable.ClassNameDeclaration;
-import net.sourceforge.pmd.lang.java.symboltable.ClassScope;
-import net.sourceforge.pmd.lang.java.symboltable.MethodNameDeclaration;
-import net.sourceforge.pmd.lang.java.symboltable.JavaNameOccurrence;
-import net.sourceforge.pmd.lang.java.symboltable.VariableNameDeclaration;
+import net.sourceforge.pmd.lang.java.symboltable.testdata.InnerClass;
+import net.sourceforge.pmd.lang.java.symboltable.testdata.InnerClass.TheInnerClass;
+import net.sourceforge.pmd.lang.java.symboltable.testdata.InnerClass.TheInnerClass.EnumTest;
 import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import net.sourceforge.pmd.lang.symboltable.NameOccurrence;
 
-import org.junit.Test;
 public class ClassScopeTest extends STBBaseTst {
 
     @Test
@@ -170,6 +169,23 @@ public class ClassScopeTest extends STBBaseTst {
         assertEquals("(String,String...)", mnd.getParameterDisplaySignature());
     }
 
+    @Test
+    public void testNestedClassesResolution() {
+        parseForClass(InnerClass.class);
+        final ASTClassOrInterfaceDeclaration n = acu.findDescendantsOfType(ASTClassOrInterfaceDeclaration.class).get(0);
+        final ClassScope c = (ClassScope) n.getScope();
+        assertEquals(InnerClass.class, c.resolveType("InnerClass"));
+        assertEquals(TheInnerClass.class, c.resolveType("InnerClass.TheInnerClass"));
+        assertEquals(TheInnerClass.class, c.resolveType("TheInnerClass")); // Within this scope, we can access it directly
+    }
+
+    @Test
+    public void testImportNestedClassesResolution() {
+        parseCode(IMPORT_NESTED_CLASSES);
+        final ASTClassOrInterfaceDeclaration n = acu.findDescendantsOfType(ASTClassOrInterfaceDeclaration.class).get(0);
+        final ClassScope c = (ClassScope) n.getScope();
+        assertEquals(EnumTest.class, c.resolveType("EnumTest"));
+    }
 
     @Test
     public final void testNestedClassDeclFound() throws Throwable {
@@ -400,6 +416,12 @@ public class ClassScopeTest extends STBBaseTst {
             "      return false;" + PMD.EOL +
             "   }" + PMD.EOL +
             "}";
+
+    private static final String IMPORT_NESTED_CLASSES =
+            "import net.sourceforge.pmd.lang.java.symboltable.testdata.InnerClass.TheInnerClass.EnumTest;" + PMD.EOL +
+            "public class Foo {" + PMD.EOL +
+            " public EnumTest e;" + PMD.EOL +
+            "}" + PMD.EOL;
 
     public static junit.framework.Test suite() {
         return new junit.framework.JUnit4TestAdapter(ClassScopeTest.class);